+++ stock/UpdateHkStockData.py
@@ -98,10 +98,11 @@ def update_all_stock_data_simple(start_date='20240410', end_date='20240412', all
         file_name = f"D:\\abu\\hk\\stock\\{code}_{all_stock_file_date}"
         stock_data.to_csv(file_name, index=False)
 
-        # # 删除文件，不删除文件可以进行每天测试
-        file_name = f"D:\\abu\\hk\\stock\\{code}_{end_date}"
-        if os.path.exists(file_name):
-            os.remove(file_name)
+        if end_date != all_stock_file_date:
+            # # 删除文件，不删除文件可以进行每天测试
+            file_name = f"D:\\abu\\hk\\stock\\{code}_{end_date}"
+            if os.path.exists(file_name):
+                os.remove(file_name)
 
 
 def pick_stock(stock_code_data, start_date='20230410', end_date='20240410'):
@@ -130,7 +131,7 @@ def pick_stock(stock_code_data, start_date='20230410', end_date='20240410'):
             choose_stock_mean.append(code)
             pick_result_df.loc[len(pick_result_df)] = [end_date, name, '\t'+code, stock_data['close'].iloc[-1], result1, result2]
 
-    pick_result_df.to_csv(f"D:\\abu\\hk\\result\\choose_hk_stock.csv", mode='a', header=True, encoding='utf-8', index=False)
+    pick_result_df.to_csv(f"D:\\abu\\hk\\result\\choose_hk_stock.csv", mode='a', header=False, encoding='utf-8', index=False)
 
 
 """
@@ -160,7 +161,7 @@ def pick_stock_ang(stock_code_data, num=10, start_date='20230410', end_date='202
             choose_stock_mean.append(code)
             pick_result_df.loc[len(pick_result_df)] = [end_date,name, '\t'+code, stock_data['close'].iloc[-1], result1, result2, result1/result2]
 
-    pick_result_df.to_csv(f"D:\\abu\\hk\\result\\choose_hk_stock_ang.csv", mode='w', header=True, encoding='utf-8',
+    pick_result_df.to_csv(f"D:\\abu\\hk\\result\\choose_hk_stock_ang.csv", mode='a', header=False, encoding='utf-8',
                           index=False)
 
 
@@ -255,7 +256,9 @@ def check_choose_stock_change(get_data_date):
         if code == 'code':
             continue
         # 从stock_prices数据框中获取该股票的最新价格
-        latest_price = stock_latest_data.loc[stock_latest_data['symbol'] == code, 'lasttrade'].values[0]
+
+        get_data = stock_latest_data.loc[stock_latest_data['symbol'] == int(code)]
+        latest_price = get_data['lasttrade'].values[0]
 
         # 从stock_history数据框中获取该股票的历史价格
         historical_price = float(check_pd.loc[check_pd['code'] == code, 'price'].values[0])
@@ -279,16 +282,16 @@ todo list
 """
 if __name__ == '__main__':
     start = time.time()
-    current_date = 20240614
+    current_date = 20240618
     # # 周一减少3天
     check_date = current_date - 1
     # 最新的数据
-    # stock_data = get_all_latest_stock()
-    # # # # 1、将每个股票的实时行情保存到历史数据，更新多天有问题,只更新一天，周一需要单独设置两个时间
-    # update_all_stock_data_simple("20230410", str(check_date), str(current_date))
-    # # # 2、对数据进行选股
-    # pick_stock(stock_data, end_date=str(current_date))
-    # pick_stock_ang(stock_data, 20, end_date=str(current_date))
+    stock_data = get_all_latest_stock()
+    # # # # # 1、将每个股票的实时行情保存到历史数据，更新多天有问题,只更新一天，周一需要单独设置两个时间
+    update_all_stock_data_simple("20230410", str(check_date), str(current_date))
+    # # # # 2、对数据进行选股
+    pick_stock(stock_data, end_date=str(current_date))
+    pick_stock_ang(stock_data, 20, end_date=str(current_date))
     # # 3、监控昨天选股情况
     check_choose_stock_change(current_date)
