+++ stock/UpdateUsStockData.py
@@ -15,7 +15,7 @@ def get_all_latest_stock():
     # 减少一天
     current_date = current_date.replace(day=current_date.day - 1)
     current_date = current_date.strftime('%Y%m%d')
-    file_name = f"D:\\abu\\us\\all\\{current_date}"
+    file_name = f"D:\\abu\\us\\all\\{current_date}.csv"
     if os.path.exists(file_name):
         stock_data = pd.read_csv(file_name)
         # stock_data.rename(columns=column_names, inplace=True)
@@ -27,7 +27,7 @@ def get_all_latest_stock():
 
 
 def pick_stock_by_date(current_date):
-    file_name = f"D:\\abu\\us\\all\\{current_date}"
+    file_name = f"D:\\abu\\us\\all\\{current_date}.csv"
     if os.path.exists(file_name):
         stock_data = pd.read_csv(file_name)
         # stock_data.rename(columns=column_names, inplace=True)
@@ -54,62 +54,6 @@ def get_stock_data_by_name(symbol, start_date='20230410', end_date='20240410'):
         return None
 
 
-# # 2. 将每个股票的实时行情保存到历史数据
-# def update_all_stock_data(start_date=20240410, end_date=20240412, single_end_data=20240412):
-#     """
-#     使用最新的实时数据来更新历史数据
-#     0、更新之前先备份一下数据
-#     1、start_date end_date 是全部最新数据文件
-#     2、更新获取最新数据
-#
-#     """
-#     # start_date = 20240410
-#     # end_date = 20240412
-#
-#     # 循环start_date到end_date
-#     cur_start_date = '20230410'
-#     for i in range(start_date, end_date):
-#         current_date = str(i + 1)
-#         stock_latest_data = pick_stock_by_date(current_date)
-#         if stock_latest_data is None:
-#             continue
-#
-#         # 将20240410 变为 2024-04-10
-#         save_current_date = f"{current_date[:4]}-{current_date[4:6]}-{current_date[6:]}"
-#         for j in stock_latest_data['代码']:
-#             # 开头是8的股票代码，不处理
-#             code = f"{j:06}"
-#             if code[0] == '8':
-#                 continue
-#
-#             end = single_end_data
-#             if single_end_data > start_date:
-#                 end = start_date
-#
-#             stock_data = get_stock_data_by_name(code, cur_start_date, str(end))
-#             if stock_data is None or stock_data.size == 0:
-#                 continue
-#
-#             # 使用concat函数将新数据添加到现有的DataFrame中
-#             stock_new_data = stock_latest_data[stock_latest_data['代码'] == j]
-#             new_data = pd.DataFrame({'date': [save_current_date], 'open': [stock_new_data['今开'].iloc[0]]
-#                                         , 'close': [stock_new_data['最新价'].iloc[0]],
-#                                      'high': [stock_new_data['最高'].iloc[0]]
-#                                         , 'low': [stock_new_data['最低'].iloc[0]],
-#                                      'volume': [stock_new_data['成交量'].iloc[0]]})
-#
-#             stock_data = pd.concat([stock_data, new_data], ignore_index=True)
-#
-#             file_name = f"D:\\abu\\us\\stock\\{code}_{cur_start_date}_{current_date}"
-#             stock_data.to_csv(file_name, index=False)
-#
-#             # 删除文件，不删除文件可以进行每天测试
-#             # file_name = f"D:\\abu\\stock\\{code}_{cur_start_date}_{start_date}"
-#             # if os.path.exists(file_name):
-#             #     os.remove(file_name)
-#         start_date = i + 1
-
-
 def update_all_stock_data_simple(start_date='20240410', end_date='20240412', all_stock_file_date='20240415'):
     """
     使用最新的实时数据来更新历史数据，只更新一天
@@ -123,18 +67,14 @@ def update_all_stock_data_simple(start_date='20240410', end_date='20240412', all
 
     # 将20240410 变为 2024-04-10
     save_current_date = f"{all_stock_file_date[:4]}-{all_stock_file_date[4:6]}-{all_stock_file_date[6:]}"
-    for j in stock_latest_data['代码']:
+    for code in stock_latest_data['代码']:
         # 开头是8的股票代码，不处理
-        code = f"{j:06}"
-        if code[0] == '8':
-            continue
-
         stock_data = get_stock_data_by_name(code, start_date, end_date)
         if stock_data is None or stock_data.size == 0:
             continue
 
         # 使用concat函数将新数据添加到现有的DataFrame中
-        stock_new_data = stock_latest_data[stock_latest_data['代码'] == j]
+        stock_new_data = stock_latest_data[stock_latest_data['代码'] == code]
         new_data = pd.DataFrame({'date': [save_current_date], 'open': [stock_new_data['开盘价'].iloc[0]]
                                     , 'close': [stock_new_data['最新价'].iloc[0]],
                                  'high': [stock_new_data['最高价'].iloc[0]]
@@ -146,16 +86,17 @@ def update_all_stock_data_simple(start_date='20240410', end_date='20240412', all
         file_name = f"D:\\abu\\us\\stock\\{code}_{start_date}_{all_stock_file_date}"
         stock_data.to_csv(file_name, index=False)
 
-        # 删除文件，不删除文件可以进行每天测试
-        file_name = f"D:\\abu\\us\\stock\\{code}_{start_date}_{end_date}"
-        if os.path.exists(file_name):
-            os.remove(file_name)
+        if end_date != all_stock_file_date:
+            # 删除文件，不删除文件可以进行每天测试
+            file_name = f"D:\\abu\\us\\stock\\{code}_{start_date}_{end_date}"
+            if os.path.exists(file_name):
+                os.remove(file_name)
 
 
 
 
-def pick_stock_new(start_date='20230410', end_date='20240410'):
-    stock_code_data = pd.read_csv('20240424')
+def pick_stock(stock_code_data, start_date='20230410', end_date='20240410'):
+    # stock_code_data = pd.read_csv('20240424')
     choose_stock_mean = []
     choose_stock_volume = []
     pick_result_df = pd.DataFrame(columns=['date', 'code', 'price', 'average', 'ang'])
@@ -164,18 +105,21 @@ def pick_stock_new(start_date='20230410', end_date='20240410'):
 
     # 循环获取
     for code in stock_code_data['代码']:
+        new_pirce = stock_code_data.loc[stock_code_data['代码'] == code, '最新价'].values[0]
+        if new_pirce < 2 or new_pirce > 500:
+            continue
         stock_data = get_stock_data_by_name(code, start_date, end_date)
-        result1 = execute_strategy_mean(stock_data, code,  start_date, end_date)
-        result2 = execute_strategy_ang(stock_data, code, start_date, end_date)
+        result1 = execute_strategy_mean(stock_data, code)
+        result2 = execute_strategy_ang(stock_data, 10, 'close')
         if result1 and result2 > 0:
             choose_stock_mean.append(code)
             pick_result_df.loc[len(pick_result_df)] = [end_date, code, stock_data['close'].iloc[-1], result1, result2]
 
-    pick_result_df.to_csv(f"D:\\abu\\us\\result\\choose_stock_mean_ang.csv", mode='a', header=True, encoding='utf-8', index=False)
+    pick_result_df.to_csv(f"D:\\abu\\us\\result\\choose_us_stock.csv", mode='a', header=True, encoding='utf-8', index=False)
 
 
-def pick_stock_2(start_date='20230410', end_date='20240410'):
-    stock_code_data = pd.read_csv('20240424')
+def pick_stock_ang(stock_code_data, start_date='20230410', end_date='20240410'):
+    # stock_code_data = pd.read_csv('20240424')
     choose_stock_mean = []
     choose_stock_volume = []
     pick_result_df = pd.DataFrame(columns=['date', 'code', 'price', 'average', 'ang'])
@@ -184,76 +128,41 @@ def pick_stock_2(start_date='20230410', end_date='20240410'):
 
     # 循环获取
     for code in stock_code_data['代码']:
+        new_pirce = stock_code_data.loc[stock_code_data['代码'] == code, '最新价'].values[0]
+        if new_pirce < 2 or new_pirce > 500:
+            continue
         stock_data = get_stock_data_by_name(code, start_date, end_date)
-        result1 = execute_strategy_ang_new(stock_data, code)
-        result2 = execute_strategy_ang(stock_data, code)
-        if result1 > 1  and result2 > 1:
+        result1 = execute_strategy_ang(stock_data, 10, 'volume')
+        result2 = execute_strategy_ang(stock_data, 10, 'close')
+        if result1 > 5  and result2 > 5:
             choose_stock_mean.append(code)
             pick_result_df.loc[len(pick_result_df)] = [end_date, code, stock_data['close'].iloc[-1], result1, result2]
 
-    pick_result_df.to_csv(f"D:\\abu\\us\\result\\choose_stock_ang.csv", mode='a', header=True, encoding='utf-8', index=False)
+    pick_result_df.to_csv(f"D:\\abu\\us\\result\\choose_us_stock_ang.csv", mode='w', header=True, encoding='utf-8', index=False)
 
 
 
-def execute_strategy_ang(pd_stock_data, code):
+def execute_strategy_ang(pd_stock_data, num, label):
     """
     这是一个成家量的选股策略
     """
     try:
         # pd_stock_data = get_stock_data_by_name(code, start_date, end_date)
-        if len(pd_stock_data) < 10:
+        if len(pd_stock_data) < num:
             return False
-        tmp = pd_stock_data[-60:]
-        ang = RegUtil.calc_regress_deg(tmp['close'], show=False)
+        tmp = pd_stock_data[-num:]
+        ang = RegUtil.calc_regress_deg(tmp[label], show=False)
         return ang
     except Exception as e:
-        print(f"获取数据失败，错误信息：{e} {code}")
+        print(f"获取数据失败，错误信息：{e} ")
         return 0
 
-def execute_strategy_ang_new(pd_stock_data, code):
-    """
-    这是一个成家量的选股策略
-    """
-    try:
-        # pd_stock_data = get_stock_data_by_name(code, start_date, end_date)
-        if len(pd_stock_data) < 10:
-            return False
-        tmp = pd_stock_data[-60:]
-        ang = RegUtil.calc_regress_deg(tmp['volume'], show=False)
-        return ang
-    except Exception as e:
-        print(f"获取数据失败，错误信息：{e} {code}")
-        return 0
 
-# def pick_stock(start_date='20230410', end_date='20240410'):
-#     stock_code_data = pd.read_csv('20240424')
-#     choose_stock_mean = []
-#     choose_stock_volume = []
-#
-#     # 循环获取
-#     for i in stock_code_data['代码']:
-#
-#         # 开头是8的股票代码，不处理
-#         result1 = execute_strategy_mean(i, start_date, end_date)
-#         if result1:
-#             choose_stock_mean.append(i)
-#
-#     with open(f"D:\\abu\\us\\result\\choose_stock_{end_date}", 'w', encoding="utf-8") as f:
-#         # f.write("--------------价格和120日均线选股策略\n")
-#         for item in choose_stock_mean:
-#             f.write(f"{item}\n")
-#         # f.write("---------------成交量选股策略\n")
-#         for item in choose_stock_volume:
-#             f.write(f"{item}\n")
-
-
-def execute_strategy_mean(stock_data, code, start_date='20230410', end_date='20240423'):
+def execute_strategy_mean(stock_data, code):
     """
     这是一个价格和120日均线的选股策略
     """
     try:
-        # stock_data = get_stock_data_by_name(code, start_date, end_date)
-
         if len(stock_data) < 120:
             return False
 
@@ -337,15 +246,16 @@ todo list
 """
 if __name__ == '__main__':
     start = time.time()
-    current_date = 20240530
-    check_date = current_date - 1
+    current_date = 20240617
+    check_date = current_date - 4
     #
     # # # 1、获取股票的实时行情
-    # get_all_latest_stock()
+    stock_data = get_all_latest_stock()
     # # # 2、将每个股票的实时行情保存到历史数据，更新多天有问题
-    # update_all_stock_data_simple("20230410", str(check_date), str(current_date))
+    update_all_stock_data_simple("20230410", str(check_date), str(current_date))
     # # 3、对数据进行选股
-    pick_stock_2(end_date=str(current_date))
+    pick_stock(stock_data, end_date=str(current_date))
+    pick_stock_ang(stock_data, end_date=str(current_date))
 
     # 4、监控昨天选股情况
     # check_choose_stock_change(20240423, 20240424)
